backend/requirements.txt
fastapi
uvicorn
requests
openai-whisper
torch
numpy
pydantic
python-multipart